requirements.txt
# requirements.txt
selenium==4.10.0
beautifulsoup4==4.12.2
webdriver-manager==3.8.6
google-generativeai==0.3.1
pandas==2.0.3
requests==2.31.0